update navigation + clean folders
