+++ src/@types/navigation.ts
@@ -1,34 +1,33 @@
 import {DrawerScreenProps} from '@react-navigation/drawer';
 import {NativeStackScreenProps} from '@react-navigation/native-stack';
 
-// main screen navigation
+export type RootNativeStackParamList = {
+  LoginStack: undefined;
+  DashboardStack: undefined;
 
-export type MainRootNativeStackParamList = {
-  Dashboard: undefined;
   MyCommunities: undefined;
+  CECTracker: undefined;
+  SkillTree: undefined;
+
+  CommunitiesStack: undefined;
+  MembersStack: {communityId: number} | undefined;
 };
 
-export type MainNativeStackScreenProp<
-  T extends keyof MainRootNativeStackParamList
-> = NativeStackScreenProps<MainRootNativeStackParamList, T>;
-
-// my communities navigation
+export type RootNativeStackScreenProp<
+  T extends keyof RootNativeStackParamList
+> = NativeStackScreenProps<RootNativeStackParamList, T>;
 
 export type RootDrawerParamList = {
-  CEC: undefined;
+  Dashboard: undefined;
+
   Communities: undefined;
-  CommunityMembers: {communityId: number} | undefined;
-  Members: undefined;
+  Members: {communityId: number} | undefined;
   Profile: undefined;
+  Report: undefined;
+
+  CEC: undefined;
+  Skills: undefined;
 };
 
 export type DrawerScreenProp<T extends keyof RootDrawerParamList> =
   DrawerScreenProps<RootDrawerParamList, T>;
-
-export type RootNativeStackParamList = {
-  CommunitiesStack: undefined;
-  CommunityMembersStack: {communityId: number} | undefined;
-};
-
-export type NativeStackScreenProp<T extends keyof RootNativeStackParamList> =
-  NativeStackScreenProps<RootNativeStackParamList, T>;
