+++ src/components/molecules/NoResult/NoResult.styles.ts
@@ -1,0 +1,14 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    padding: 50,
+  },
+  text: {
+    fontSize: 20,
+  },
+});
+
+export default styles;
