+++ src/components/molecules/index.ts
@@ -3,4 +3,5 @@ export * from './CommunityDetailsModal';
 export * from './DrawerAccordion';
 export * from './LabeledInput';
 export * from './MemberProgressChart';
+export * from './NoResult';
 export * from './Search';
