--- src/components/screens/CEC/CECDashboardScreen/index.tsx
@@ -1,5 +1,0 @@
-import {CECDashboardTemplate} from 'components/templates/CEC';
-
-export const CECDashboardScreen = () => {
-  return <CECDashboardTemplate />;
-};
