+++ src/components/screens/CECTracker/CECDashboardScreen/index.tsx
@@ -1,0 +1,5 @@
+import {CECDashboardTemplate} from 'components/templates/CECTracker';
+
+export const CECDashboardScreen = () => {
+  return <CECDashboardTemplate />;
+};
