--- src/components/screens/CommunityMembersScreen/index.tsx
@@ -1,8 +1,0 @@
-import {useCommunitiesDataProvider} from 'providers/CommunitiesDataProvider';
-import {CommunityMembersTemplate} from 'components/templates';
-
-export const CommunityMembersScreen = () => {
-  const {communityList} = useCommunitiesDataProvider();
-
-  return <CommunityMembersTemplate communityList={communityList} />;
-};
