+++ src/components/screens/DashboardScreen/index.tsx
@@ -1,40 +1,38 @@
 import {useNavigation} from '@react-navigation/native';
 import {NativeStackNavigationProp} from '@react-navigation/native-stack';
 
-import {MainRootNativeStackParamList} from '../../../@types/navigation';
+import {RootNativeStackParamList} from '../../../@types/navigation';
 import {AppCardObject} from 'components/molecules';
 import {DashboardTemplate} from 'components/templates';
 import {COLORS} from 'constants/colors';
+import {StackScreen} from 'constants/navigation';
 import {useUserDataProvider} from 'providers/UserDataProvider';
-import {MainStackScreenName} from 'constants/enums';
-
-type DashboardStackNavigationProp =
-  NativeStackNavigationProp<MainRootNativeStackParamList>;
 
 export const DashboardScreen = () => {
   const {user} = useUserDataProvider();
-  const {navigate} = useNavigation<DashboardStackNavigationProp>();
+  const {navigate} =
+    useNavigation<NativeStackNavigationProp<RootNativeStackParamList>>();
 
   const applications: AppCardObject[] = [
     {
       title: 'My Community',
       icon: {
-        name: 'people',
+        name: 'groups',
         type: 'material',
         size: 50,
         color: COLORS.ULTRA_LIGHT_GRAY,
       },
-      onPress: () => navigate(MainStackScreenName.MyCommunities),
+      onPress: () => navigate(StackScreen.MyCommunities),
     },
     {
       title: 'CEC Tracker',
       icon: {
-        name: 'notebook',
-        type: 'simple-line-icon',
+        name: 'assignment',
+        type: 'material',
         size: 50,
         color: COLORS.ULTRA_LIGHT_GRAY,
       },
-      onPress: () => console.log('CEC Tracker'),
+      onPress: () => navigate(StackScreen.CECTracker),
     },
     {
       title: 'Skill Tree',
@@ -44,7 +42,7 @@ export const DashboardScreen = () => {
         size: 50,
         color: COLORS.ULTRA_LIGHT_GRAY,
       },
-      onPress: () => console.log('Skill Tree'),
+      onPress: () => navigate(StackScreen.SkillTree),
     },
   ];
