+++ src/components/screens/LoginScreen/index.tsx
@@ -1,0 +1,5 @@
+import {LoginTemplate} from 'components/templates';
+
+export const LoginScreen = () => {
+  return <LoginTemplate />;
+};
