--- src/components/screens/MembersScreen/index.tsx
@@ -1,5 +1,0 @@
-import {MembersTemplate} from 'components/templates';
-
-export const MembersScreen = () => {
-  return <MembersTemplate />;
-};
