+++ src/components/screens/MyCommunities/MembersScreen/index.tsx
@@ -1,0 +1,8 @@
+import {useCommunitiesDataProvider} from 'providers/CommunitiesDataProvider';
+import {MembersTemplate} from 'components/templates';
+
+export const MembersScreen = () => {
+  const {communityList} = useCommunitiesDataProvider();
+
+  return <MembersTemplate communityList={communityList} />;
+};
