+++ src/components/screens/MyCommunities/ReportScreen/index.tsx
@@ -1,0 +1,5 @@
+import {ReportTemplate} from 'components/templates';
+
+export const ReportScreen = () => {
+  return <ReportTemplate />;
+};
