+++ src/components/screens/MyCommunities/index.ts
@@ -1,0 +1,4 @@
+export * from './CommunitiesDashboardScreen';
+export * from './MembersScreen';
+export * from './ProfileScreen';
+export * from './ReportScreen';
