+++ src/components/screens/SkillTree/SkillsDashboardScreen/index.tsx
@@ -1,0 +1,5 @@
+import {SkillsDashboardTemplate} from 'components/templates';
+
+export const SkillsDashboardScreen = () => {
+  return <SkillsDashboardTemplate />;
+};
