+++ src/components/screens/index.ts
@@ -1,6 +1,5 @@
-export * from './CEC';
-export * from './CommunitiesDashboardScreen';
-export * from './CommunityMembersScreen';
+export * from './CECTracker';
 export * from './DashboardScreen';
-export * from './MembersScreen';
-export * from './ProfileScreen';
+export * from './LoginScreen';
+export * from './MyCommunities';
+export * from './SkillTree';
