--- src/components/templates/CEC/CECDashboardTemplate/index.tsx
@@ -1,10 +1,0 @@
-import {AppContainer, Text} from 'components/atoms';
-import styles from './CECDashboardTemplate.styles';
-
-export const CECDashboardTemplate = () => {
-  return (
-    <AppContainer style={styles.container}>
-      <Text style={styles.text}>CEC Requests</Text>
-    </AppContainer>
-  );
-};
