--- src/components/templates/CommunityMembersTemplate/CommunityMembersTemplate.styles.tsx
@@ -1,20 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-const styles = StyleSheet.create({
-  search: {
-    paddingBottom: 10,
-  },
-  noResultContainer: {
-    flex: 1,
-    alignItems: 'center',
-    padding: 50,
-  },
-  noResultText: {
-    fontSize: 20,
-  },
-  listContentContainer: {
-    paddingVertical: 5,
-  },
-});
-
-export default styles;
