--- src/components/templates/CommunityTrackerDashboardTemplate/index.tsx
@@ -1,11 +1,0 @@
-import {View} from 'react-native';
-
-import {Text} from 'components/atoms';
-
-export const CommunityTrackerDashboardTemplate = () => {
-  return (
-    <View>
-      <Text>Community Tracker Dashboard</Text>
-    </View>
-  );
-};
