+++ src/components/templates/LoginTemplate/index.tsx
@@ -1,0 +1,9 @@
+import {AppContainer, Text} from 'components/atoms';
+
+export const LoginTemplate = () => {
+  return (
+    <AppContainer keyboardShouldPersistTaps="handled">
+      <Text>Screen is still not available</Text>
+    </AppContainer>
+  );
+};
