--- src/components/templates/MembersTemplate/MembersTemplate.styles.tsx
@@ -1,9 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-const styles = StyleSheet.create({
-  text: {
-    fontSize: 24,
-  },
-});
-
-export default styles;
