+++ src/components/templates/MyCommunities/CommunitiesDashboardTemplate/CommunitiesDashboardTemplate.styles.tsx
