+++ src/components/templates/MyCommunities/MembersTemplate/MembersTemplate.styles.tsx
@@ -1,0 +1,12 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  search: {
+    paddingBottom: 10,
+  },
+  listContentContainer: {
+    paddingVertical: 5,
+  },
+});
+
+export default styles;
