+++ src/components/templates/MyCommunities/index.ts
@@ -1,0 +1,4 @@
+export * from './CommunitiesDashboardTemplate';
+export * from './MembersTemplate';
+export * from './ProfileTemplate';
+export * from './ReportTemplate';
