+++ src/components/templates/SkillTree/index.ts
@@ -1,0 +1,1 @@
+export * from './SkillsDashboardTemplate';
