+++ src/components/templates/index.ts
@@ -1,6 +1,5 @@
-export * from './CommunitiesDashboardTemplate';
-export * from './CommunityMembersTemplate';
-export * from './CommunityTrackerDashboardTemplate';
+export * from './CECTracker';
 export * from './DashboardTemplate';
-export * from './MembersTemplate';
-export * from './ProfileTemplate';
+export * from './LoginTemplate';
+export * from './MyCommunities';
+export * from './SkillTree';
