+++ src/constants/enums.ts
@@ -4,31 +4,3 @@ export enum HTTPMethod {
   PUT = 'PUT',
   DELETE = 'DELETE',
 }
-
-export enum MainStackScreenName {
-  Dashboard = 'Dashboard',
-  MyCommunities = 'MyCommunities',
-}
-
-// My Communities Screen
-
-export enum ScreenName {
-  CEC = 'CEC',
-  Communities = 'Communities',
-  CommunityMembers = 'CommunityMembers',
-  Members = 'Members',
-  MembersDetails = 'MembersDetails',
-  Profile = 'Profile',
-  Report = 'Report',
-}
-
-export enum StackScreenName {
-  CommunitiesStack = 'CommunitiesStack',
-  CommunityMembersStack = 'CommunityMembersStack',
-}
-
-export enum ScreenTitle {
-  Communities = 'Communities',
-  Members = 'Members',
-  Profile = 'Profile',
-}
