+++ src/constants/navigation.ts
@@ -1,0 +1,34 @@
+export enum DrawerScreen {
+  Dashboard = 'Dashboard',
+
+  Communities = 'Communities',
+  Members = 'Members',
+  Profile = 'Profile',
+  Report = 'Report',
+
+  CEC = 'CEC',
+  Skills = 'Skills',
+}
+
+export enum StackScreen {
+  LoginStack = 'LoginStack',
+  DashboardStack = 'DashboardStack',
+
+  MyCommunities = 'MyCommunities',
+  CECTracker = 'CECTracker',
+  SkillTree = 'SkillTree',
+
+  CommunitiesStack = 'CommunitiesStack',
+  MembersStack = 'MembersStack',
+}
+
+export enum ScreenTitle {
+  Communities = 'Communities',
+  Members = 'Members',
+  Profile = 'Profile',
+  Report = 'Report',
+
+  CEC = 'CEC',
+
+  Skills = 'Skills',
+}
